feat: introduce email builder utilities for UIBuilder integration
