+++ app/docs/docs-data/docs-page-layers/page-type-renderers.ts
@@ -37,7 +37,7 @@ export const PAGE_TYPE_RENDERERS_LAYER = {
         "type": "Markdown",
         "name": "Markdown",
         "props": {},
-        "children": "## Email Builder Example\n\nThe following example wires up UI Builder with react-email component support. UI Builder does **not** import `@react-email/*` — all email-specific logic lives in consumer code.\n\n### 1. Install react-email components registry\n\n```bash\nnpx shadcn@latest add https://raw.githubusercontent.com/olliethedev/ui-builder/main/registry/react-email-components-registry.json\n```\n\nThis installs `lib/ui-builder/registry/react-email-component-definitions.ts` and the required npm packages (`@react-email/components`, `@react-email/render`, `web-streams-polyfill`).\n\n### 2. Wire up the email builder\n\nReact-email components like `<Html>`, `<Head>`, and `<Body>` render actual structural HTML elements. When placed inside the editor's canvas iframe these elements are suppressed by the browser, producing a blank canvas. The solution is a **canvas-specific registry** that swaps those structural components with rendering-safe div/null substitutes — while keeping the real `@react-email` components in the registry for code generation.\n\n```tsx\nimport { useMemo } from 'react';\nimport UIBuilder from '@/components/ui/ui-builder';\nimport LayerRenderer from '@/components/ui/ui-builder/layer-renderer';\nimport { reactEmailComponentDefinitions } from '@/lib/ui-builder/registry/react-email-component-definitions';\nimport type {\n  PageTypeRenderer,\n  PageTypeCodeGenerator,\n  PageTypeRendererProps,\n  ComponentLayer,\n  ComponentRegistry,\n} from '@/components/ui/ui-builder/types';\n\n// Canvas-safe substitutes for structural HTML elements.\n// The real @react-email components are preserved in the registry for code generation.\nconst NullComponent = () => null;\nconst CanvasHtml = ({ children }: { children?: React.ReactNode }) => (\n  <div data-email-html style={{ display: 'contents' }}>{children}</div>\n);\nconst CanvasBody = ({ children, className }: { children?: React.ReactNode; className?: string }) => (\n  <div data-email-body className={className} style={{ backgroundColor: 'white', minHeight: '100%' }}>{children}</div>\n);\n\n// Canvas renderer — swaps structural HTML elements with div substitutes\n// so the email content is visible and interactive in the editor canvas.\nfunction EmailCanvasRenderer({ page, componentRegistry, editorConfig }: PageTypeRendererProps) {\n  const canvasRegistry: ComponentRegistry = useMemo(() => {\n    const base = { ...componentRegistry };\n    if (base.Html) base.Html = { ...base.Html, component: CanvasHtml };\n    if (base.Head) base.Head = { ...base.Head, component: NullComponent };\n    if (base.Preview) base.Preview = { ...base.Preview, component: NullComponent };\n    if (base.Body) base.Body = { ...base.Body, component: CanvasBody };\n    return base;\n  }, [componentRegistry]);\n\n  return (\n    <LayerRenderer\n      className=\"contents\"\n      page={page}\n      componentRegistry={canvasRegistry}\n      editorConfig={editorConfig}\n    />\n  );\n}\n\nconst emailPageRenderer: PageTypeRenderer = {\n  label: 'Email',\n  defaultRootLayerType: 'Html',        // new email pages start with <Html> root\n  defaultRootLayerProps: { lang: 'en' },\n  renderEditorCanvas: (props) => <EmailCanvasRenderer {...props} />,\n};\n\n// Code generator — produces JSX + usage example for the email template.\n// Uses the original registry (with real @react-email components) for accurate imports.\nconst emailCodeGenerator: PageTypeCodeGenerator = {\n  label: 'Email JSX',\n  generateCode: (page: ComponentLayer, registry: ComponentRegistry) => {\n    // Build import statements and JSX from the layer tree\n    // See app/platform/email-builder.tsx for a full implementation\n    return `// Generated email template for ${page.name}`;\n  },\n};\n\nexport function EmailBuilder() {\n  return (\n    <UIBuilder\n      componentRegistry={reactEmailComponentDefinitions}\n      pageTypeRenderers={{ email: emailPageRenderer }}\n      pageTypeCodeGenerators={{ email: emailCodeGenerator }}\n      allowPagesCreation\n    />\n  );\n}\n```\n\n### 3. Safari / iOS compatibility\n\nAdd this import at the **top** of your email route's entry file for Safari/iOS compatibility with `@react-email/render`:\n\n```ts\nimport 'web-streams-polyfill/polyfill';\n```"
+        "children": "## Email Builder Example\n\nThe react-email registry ships everything needed to wire up UI Builder as a full email editor — component definitions, canvas substitutes, code generator, and page renderer config. UI Builder itself has zero knowledge of `@react-email/*`.\n\n### 1. Install react-email components registry\n\n```bash\nnpx shadcn@latest add https://raw.githubusercontent.com/olliethedev/ui-builder/main/registry/react-email-components-registry.json\n```\n\nThis installs:\n- `lib/ui-builder/registry/react-email-component-definitions.ts` — `ComponentRegistry` for all 14 react-email components\n- `lib/ui-builder/email/email-builder-utils.tsx` — ready-to-use `emailPageRenderer`, `emailCodeGenerator`, and canvas substitutes\n- npm packages: `@react-email/components`, `@react-email/render`, `web-streams-polyfill`\n\n### 2. Wire up the email builder\n\nImport and pass the pre-built config objects to `UIBuilder`:\n\n```tsx\nimport UIBuilder from '@/components/ui/ui-builder';\nimport { reactEmailComponentDefinitions } from '@/lib/ui-builder/registry/react-email-component-definitions';\nimport {\n  emailPageRenderer,\n  emailCodeGenerator,\n} from '@/lib/ui-builder/email/email-builder-utils';\n\nexport function EmailBuilder() {\n  return (\n    <UIBuilder\n      componentRegistry={reactEmailComponentDefinitions}\n      pageTypeRenderers={{ email: emailPageRenderer }}\n      pageTypeCodeGenerators={{ email: emailCodeGenerator }}\n      allowPagesCreation\n    />\n  );\n}\n```\n\n### 3. Safari / iOS compatibility\n\nAdd this import at the **top** of your email route's entry file for Safari/iOS compatibility with `@react-email/render`:\n\n```ts\nimport 'web-streams-polyfill/polyfill';\n```\n\n### Customizing the Email Builder\n\nThe `email-builder-utils.tsx` file also exports the lower-level building blocks if you need to customize behavior — for example, to modify how structural HTML elements are substituted on the canvas:\n\n```tsx\nimport { useMemo } from 'react';\nimport LayerRenderer from '@/components/ui/ui-builder/layer-renderer';\nimport {\n  NullComponent,\n  CanvasHtml,\n  CanvasBody,\n  generateEmailCode,\n} from '@/lib/ui-builder/email/email-builder-utils';\nimport type {\n  PageTypeRenderer,\n  PageTypeRendererProps,\n  ComponentRegistry,\n} from '@/components/ui/ui-builder/types';\n\n// Custom canvas renderer — add your own substitutes or logic\nfunction MyEmailCanvasRenderer({ page, componentRegistry, editorConfig }: PageTypeRendererProps) {\n  const canvasRegistry: ComponentRegistry = useMemo(() => {\n    const base = { ...componentRegistry };\n    if (base.Html) base.Html = { ...base.Html, component: CanvasHtml };\n    if (base.Head) base.Head = { ...base.Head, component: NullComponent };\n    if (base.Preview) base.Preview = { ...base.Preview, component: NullComponent };\n    if (base.Body) base.Body = { ...base.Body, component: CanvasBody };\n    return base;\n  }, [componentRegistry]);\n\n  return (\n    <LayerRenderer\n      className=\"contents\"\n      page={page}\n      componentRegistry={canvasRegistry}\n      editorConfig={editorConfig}\n    />\n  );\n}\n\nconst myEmailPageRenderer: PageTypeRenderer = {\n  label: 'Email',\n  defaultRootLayerType: 'Html',\n  defaultRootLayerProps: { lang: 'en' },\n  renderEditorCanvas: (props) => <MyEmailCanvasRenderer {...props} />,\n};\n```"
       },
       {
         "id": "page-type-renderers-email-demo",
